fixed init command bug

+++ cmd/init.go
@@ -18,11 +18,14 @@ var initCmd = &cobra.Command{
 		fmt.Print("Initialize ok(Y/n) > ")
 		fmt.Scan(&stdin)
 		if stdin == "Y" {
-			if err := db.DeleteDbFile(); err != nil {
-				return err
-			}
-			if err := memoRepository.DeleteAllMemo(); err != nil {
-				return err
+
+			if err := db.ExistsDbFile(); err == nil {
+				if err := db.DeleteDbFile(); err != nil {
+					return err
+				}
+				if err := memoRepository.DeleteAllMemo(); err != nil {
+					return err
+				}
 			}
 			db.InitDb()
 			fmt.Println("Initialized taskman.")
